[core] Add default handling for time parameters
Instead of using the same logic across all plugins to parse the time
parameters from the current url, we are moving this function to the core
package, so that we can use it from there in all plugins.

This also fixes a small bug, which was introduced in the previous
commit. There we added a sorting option for the times column in the
ClickHouse plugin, which showed the wrong sort order for the times
column. This should be fixed now.
