+++ plugins/istio/src/utils/helpers.ts
@@ -1,44 +1,21 @@
-import { IPluginTimes, TTime, TTimeOptions } from '@kobsio/plugin-core';
+import { IPluginTimes, getTimeParams } from '@kobsio/plugin-core';
 import { IOptions } from './interfaces';
 
 // getApplicationsOptionsFromSearch is used to get the Istio options from a given search location.
 export const getApplicationsOptionsFromSearch = (search: string): IOptions => {
   const params = new URLSearchParams(search);
   const namespaces = params.getAll('namespace');
-  const time = params.get('time');
-  const timeEnd = params.get('timeEnd');
-  const timeStart = params.get('timeStart');
 
   return {
     namespaces: namespaces.length > 0 ? namespaces : [],
-    times: {
-      time: time && TTimeOptions.includes(time) ? (time as TTime) : 'last15Minutes',
-      timeEnd:
-        time && TTimeOptions.includes(time) && timeEnd ? parseInt(timeEnd as string) : Math.floor(Date.now() / 1000),
-      timeStart:
-        time && TTimeOptions.includes(time) && timeStart
-          ? parseInt(timeStart as string)
-          : Math.floor(Date.now() / 1000) - 900,
-    },
+    times: getTimeParams(params),
   };
 };
 
 // getApplicationOptionsFromSearch is used to get the Istio options from a given search location.
 export const getApplicationOptionsFromSearch = (search: string): IPluginTimes => {
   const params = new URLSearchParams(search);
-  const time = params.get('time');
-  const timeEnd = params.get('timeEnd');
-  const timeStart = params.get('timeStart');
-
-  return {
-    time: time && TTimeOptions.includes(time) ? (time as TTime) : 'last15Minutes',
-    timeEnd:
-      time && TTimeOptions.includes(time) && timeEnd ? parseInt(timeEnd as string) : Math.floor(Date.now() / 1000),
-    timeStart:
-      time && TTimeOptions.includes(time) && timeStart
-        ? parseInt(timeStart as string)
-        : Math.floor(Date.now() / 1000) - 900,
-  };
+  return getTimeParams(params);
 };
 
 // formatNumber brings the given value returned by Prometheus as string into a format we can use in our ui.
